Beta/rpcvr 1.5

+++ MainForm.cs
@@ -410,6 +410,7 @@ namespace RookiePCVR
         public static int updint = 0;
         public static bool nodeviceonstart = false;
         public static bool either = false;
+        private bool _allItemsInitialized = false;
 
         private async void initListPCVRView()
         {
@@ -461,6 +462,11 @@ namespace RookiePCVR
             gamesListView.Items.Clear();
             gamesListView.Items.AddRange(arr);
             gamesListView.EndUpdate();
+            if (!_allItemsInitialized)
+            {
+                _allItems = gamesListView.Items.Cast<ListViewItem>().ToList();
+                _allItemsInitialized = true; // Set the flag to true after initialization
+            }
             ChangeTitle("                                                \n\n");
             loaded = true;
         }
@@ -612,6 +618,7 @@ Things you can try:
                     gameName = gamesQueueList.ToArray()[0];
                     string dir = Path.GetDirectoryName(gameName);
                     string gameDirectory = Properties.Settings.Default.downloadDir + "\\" + gameName;
+                    string downloadDirectory = Path.Combine(Properties.Settings.Default.downloadDir, gameName);
                     string path = gameDirectory;
 
                     string gameNameHash = string.Empty;
@@ -662,12 +669,13 @@ Things you can try:
 
                         if (doDownload)
                         {
+                            downloadDirectory = $"{Properties.Settings.Default.downloadDir}\\{gameNameHash}";
                             _ = Logger.Log($"rclone copy \"Public:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\"");
                             t1 = new Thread(() =>
                             {
                                 string rclonecommand =
-                                $"copy \":http:/{gameNameHash}/\" \"{Properties.Settings.Default.downloadDir}\\{gameNameHash}\" {extraArgs} {virtualFilesystemCompatibilityArg} --progress --rc";
-                                gameDownloadOutput = RCLONE.runRcloneCommand_PublicConfig(rclonecommand);
+                                $"copy \":http:/{gameNameHash}/\" \"{downloadDirectory}\" {extraArgs} {virtualFilesystemCompatibilityArg} --progress --rc --check-first --fast-list";
+                                gameDownloadOutput = RCLONE.runRcloneCommand_PublicConfig(rclonecommand, true);
                             });
                         }
                         else
@@ -678,11 +686,22 @@ Things you can try:
                     else
                     {
                         _ = Directory.CreateDirectory(gameDirectory);
-                        _ = Logger.Log($"rclone copy \"{currentRemote}:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\"");
+                        downloadDirectory = $"{SideloaderRCLONE.RcloneGamesFolder}/{gameName}";
+                        _ = Logger.Log($"rclone copy \"{currentRemote}:{downloadDirectory}\"");
                         t1 = new Thread(() =>
                         {
-                            gameDownloadOutput = RCLONE.runRcloneCommand_DownloadConfig($"copy \"{currentRemote}:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\" \"{Properties.Settings.Default.downloadDir}\\{gameName}\" {extraArgs} {virtualFilesystemCompatibilityArg} --progress --rc --retries 1 --low-level-retries 1");
+                            gameDownloadOutput = RCLONE.runRcloneCommand_DownloadConfig($"copy \"{currentRemote}:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\" \"{Properties.Settings.Default.downloadDir}\\{gameName}\" {extraArgs} {virtualFilesystemCompatibilityArg} --progress --rc --retries 1 --low-level-retries 1 --check-first");
                         });
+                    }
+
+                    if (Directory.Exists(downloadDirectory))
+                    {
+                        string[] partialFiles = Directory.GetFiles($"{downloadDirectory}", "*.partial");
+                        foreach (string file in partialFiles)
+                        {
+                            File.Delete(file);
+                            Logger.Log($"Deleted partial file: {file}");
+                        }
                     }
 
                     t1.IsBackground = true;
@@ -690,7 +709,7 @@ Things you can try:
 
                     ChangeTitle("Downloading game " + gameName, false);
                     speedLabel.Text = "Starting download..."; etaLabel.Text = "Please wait...";
-                    int i = 0;
+
                     //Download
                     while (t1.IsAlive)
                     {
@@ -698,42 +717,55 @@ Things you can try:
                         {
                             HttpResponseMessage response = await client.PostAsync("http://127.0.0.1:5572/core/stats", null);
                             string foo = await response.Content.ReadAsStringAsync();
-                            Debug.WriteLine("RESP CONTENT " + foo);
+                            //Debug.WriteLine("RESP CONTENT " + foo);
                             dynamic results = JsonConvert.DeserializeObject<dynamic>(foo);
-
+                            
                             if (results["transferring"] != null)
                             {
-                                long allSize = 0;
-                                long downloaded = 0;
+                                double totalSize = 0;
+                                double downloadedSize = 0;
+                                long fileCount = 0;
+                                long transfersComplete = 0;
+                                long totalChecks = 0;
+                                long globalEta = 0;
+                                float speed = 0;
+                                float downloadSpeed = 0;
+                                double estimatedFileCount = 0;
 
-                                foreach (dynamic obj in results.transferring)
-                                {
-                                    allSize += obj["size"].ToObject<long>();
-                                    downloaded += obj["bytes"].ToObject<long>();
+                                totalSize = results["totalBytes"];
+                                downloadedSize = results["bytes"];
+                                fileCount = results["totalTransfers"];
+                                totalChecks = results["totalChecks"];
+                                transfersComplete = results["transfers"];
+                                globalEta = results["eta"];
+                                speed = results["speed"];
+                                estimatedFileCount = Math.Ceiling(totalSize / 524288000); // maximum part size
+
+                                if (totalChecks > fileCount) {
+                                    fileCount = totalChecks;
+                                }
+                                if (estimatedFileCount > fileCount) {
+                                    fileCount = (long)estimatedFileCount;
                                 }
 
-                                float downloadSpeed = results.speed.ToObject<float>() / 1000000;
-                                allSize /= 1000000;
-                                downloaded /= 1000000;
+                                downloadSpeed = speed / 1000000;
+                                totalSize /= 1000000;
+                                downloadedSize /= 1000000;
 
-                                Debug.WriteLine("Allsize: " + allSize + "\nDownloaded: " + downloaded + "\nValue: " + (downloaded / (double)allSize * 100));
+                                // Logger.Log("Files: " + transfersComplete.ToString() + "/" + fileCount.ToString() + " (" + Convert.ToInt32((downloadedSize / totalSize) * 100).ToString() + "% Complete)");
+                                // Logger.Log("Downloaded: " + downloadedSize.ToString() + " of " + totalSize.ToString());
 
                                 progressBar.Style = ProgressBarStyle.Continuous;
-                                progressBar.Value = Convert.ToInt32(downloaded / (double)allSize * 100);
+                                progressBar.Value = Convert.ToInt32((downloadedSize / totalSize) * 100);
 
-                                i++;
-                                if (i == 4)
-                                {
-                                    i = 0;
-                                    float seconds = (allSize - downloaded) / downloadSpeed;
-                                    TimeSpan time = TimeSpan.FromSeconds(seconds);
-                                    etaLabel.Text = "ETA: " + time.ToString(@"hh\:mm\:ss") + " left";
-                                }
+                                TimeSpan time = TimeSpan.FromSeconds(globalEta);
+                                etaLabel.Text = etaLabel.Text = "ETA: " + time.ToString(@"hh\:mm\:ss") + " left";
 
-                                speedLabel.Text = "DLS: " + string.Format("{0:0.00}", downloadSpeed) + " MB/s";
+                                speedLabel.Text = "DLS: " + transfersComplete.ToString() + "/" + fileCount.ToString() + " files - " + string.Format("{0:0.00}", downloadSpeed) + " MB/s";
                             }
                         }
-                        catch { }
+                        catch {
+                        }
 
                         await Task.Delay(100);
 
@@ -745,21 +777,28 @@ Things you can try:
                         try
                         {
                             cleanupActiveDownloadStatus();
-                            if (hasPublicPCVRConfig)
-                            {
-                                if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}"))
-                                {
-                                    Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}", true);
-                                }
 
-                                if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameName}"))
+                            DialogResult res = FlexibleMessageBox.Show(
+                                $"{gameName} already has some downloaded files, do you want to delete them?\n\nClick NO to keep the files if you wish to resume your download later.",
+                                "Delete Temporary Files?", MessageBoxButtons.YesNo);
+
+                            if (res == DialogResult.Yes) {
+                                if (hasPublicPCVRConfig)
                                 {
-                                    Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameName}", true);
+                                    if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}"))
+                                    {
+                                        Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}", true);
+                                    }
+
+                                    if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameName}"))
+                                    {
+                                        Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameName}", true);
+                                    }
                                 }
-                            }
-                            else
-                            {
-                                Directory.Delete(Properties.Settings.Default.downloadDir + "\\" + gameName, true);
+                                else
+                                {
+                                    Directory.Delete(Properties.Settings.Default.downloadDir + "\\" + gameName, true);
+                                }
                             }
                         }
                         catch (Exception ex)
@@ -980,29 +1019,26 @@ Things you can try:
             // Cancel any ongoing searches
             _cts?.Cancel();
 
-            _allItems = gamesListView.Items.Cast<ListViewItem>().ToList();
-
             string searchTerm = searchTextBox.Text;
             if (!string.IsNullOrEmpty(searchTerm))
             {
                 _cts = new CancellationTokenSource();
+
                 try
                 {
-                    var matches = await Task.Run(() =>
-                        _allItems
-                            .Where(i => i.Text.IndexOf(searchTerm, StringComparison.CurrentCultureIgnoreCase) >= 0)
-                            .ToList(),
-                        _cts.Token);
+                    var matches = _allItems
+                        .Where(i => i.Text.IndexOf(searchTerm, StringComparison.CurrentCultureIgnoreCase) >= 0)
+                        .ToList();
 
-                    // Update UI on UI thread
-                    Invoke(new Action(() =>
+                    gamesListView.BeginUpdate(); // Improve UI performance
+                    gamesListView.Items.Clear();
+
+                    foreach (var match in matches)
                     {
-                        gamesListView.Items.Clear();
-                        foreach (var match in matches)
-                        {
-                            gamesListView.Items.Add(match);
-                        }
-                    }));
+                        gamesListView.Items.Add(match);
+                    }
+
+                    gamesListView.EndUpdate(); // End the update to refresh the UI
                 }
                 catch (OperationCanceledException)
                 {
@@ -1011,7 +1047,6 @@ Things you can try:
             }
             else
             {
-                // No matching items found, restore the original list
                 initListPCVRView();
             }
         }
